STY: First pass at eliminating compiler warnings.
There might be more I wanted to add, but I don't remember.
I'll have to add those the way I did the first time,
compiling with -Werror until it works.
This pass is almost exclusively focused on the registry.
I'll have a few more for Fortran once I figure out how.

Inspiration: registry was segfaulting, so I tried fixing the warnings
in case it was one of those.  It didn't work, but it's worth doing
regardless.

+++ tools/registry.c
@@ -1,16 +1,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 #ifdef _WIN32
+#define rindex(X,Y) strrchr(X,Y)
+#define index(X,Y) strchr(X,Y)
+#endif
 # include <io.h>
-# define rindex(X,Y) strrchr(X,Y)
-# define index(X,Y) strchr(X,Y)
-#else
 # include <sys/time.h>
 # include <sys/resource.h>
 # include <unistd.h>
 # include <string.h>
 # include <strings.h>
-#endif
 
 #define DEFINE_GLOBALS
 #include "protos.h"
@@ -23,13 +22,14 @@
 
 /* SamT: bug fix: main returns int */
 int
-main( int argc, char *argv[], char *env[] )
+main( int argc, char *argv[] )
 {
-  char fname_in[NAMELEN], dir[NAMELEN], fname_tmp[NAMELEN], command[NAMELEN] ;
-  char fname_wrk[NAMELEN] ;
+  char fname_in[NAMELEN] = {'\0'}, dir[NAMELEN] = {'\0'};
+  char fname_tmp[NAMELEN] = {'\0'}, command[2 * NAMELEN + 32] = {'\0'};
+  char fname_wrk[NAMELEN + 18] = {'\0'};
   FILE * fp_in, *fp_tmp ;
-  char * thisprog  ;
-  char *env_val ;
+  char * thisprog = "";
+  char *env_val = "";
   int mypid ;
   int do_irr_diag ;
 #ifndef _WIN32
@@ -37,7 +37,7 @@ main( int argc, char *argv[], char *env[] )
 #endif
 
   mypid = (int) getpid() ;
-  strcpy( thiscom, argv[0] ) ;
+  strncpy( thiscom, argv[0], 4 * NAMELEN - 1) ;
   argv++ ;
 
   sw_deref_kludge           = 0 ;
@@ -160,7 +160,23 @@ main( int argc, char *argv[], char *env[] )
       sprintf( fname_wrk,"%s/Registry_irr_diag",dir ) ;
     }
 //  fprintf(stderr,"Registry tmp file = %s\n",fname_wrk);
-    sprintf(command,"/bin/cp %s %s\n",fname_in,fname_wrk);
+    /* we should be able to implement this using posix_spawn */
+    /*
+      #include <spawn.h>
+      extern char **environ;
+      pid_t child_pid;
+      // There doesn't seem to be a way to specify the length, so I'm
+      // assuming it's null-terminated
+      char *command_argv[4] = {"/bin/cp", NULL, NULL, NULL};
+      command_argv[1] = fname_in;
+      command_argv[2] = fname_wrk;
+
+      if (posix_spawn(&child_pid, command_argv[0], NULL, NULL, command_argv, environ)) {
+        fprintf(stderr, "Could not copy %s to %s\n", fname_in, fname_wrk);
+	exit(2);
+      }
+     */
+    sprintf(command,"/bin/cp \'%s\' \'%s\'\n",fname_in,fname_wrk);
 //  fprintf(stderr,"Command = %s\n",command);
     if( system( command ) ) {
       fprintf(stderr,"Could not copy %s to %s\n",fname_in,fname_wrk);
@@ -172,6 +188,15 @@ main( int argc, char *argv[], char *env[] )
       exit(2) ;
     }
     if( !access( "Registry/registry.irr_diag",F_OK ) ) {
+      /*
+	command_argv[0] = "/bin/rm";
+	command_argv[1] = "-f";
+	command_argv[2] = "Registry/registry.irr_diag";
+	if (posix_spawn(&child_pid, command_argv[0], NULL, NULL, command_argv, environ)) {
+          fprintf(stderr, "Could not remove Registry/registry.irr_diag\n", fname_in, fname_wrk);
+	  exit(2);
+	}
+      */
       sprintf(command,"/bin/rm -f Registry/registry.irr_diag\n");
       if( system( command ) ) {
         fprintf(stderr,"Could not remove Registry/registry.irr_diag\n");
@@ -282,6 +307,15 @@ cleanup:
    sprintf(command,"del /F /Q %s\n",fname_tmp );
 #else
    if( do_irr_diag ) {
+      /*
+	command_argv[0] = "/bin/rm";
+	command_argv[1] = "-f";
+	command_argv[2] = fname_wrk;
+	if (posix_spawn(&child_pid, command_argv[0], NULL, NULL, command_argv, environ)) {
+          fprintf(stderr, "Could not remove %s\n", fname_wrk);
+	  exit(2);
+	}
+      */
      sprintf(command,"/bin/rm -f %s\n",fname_wrk );
      system( command ) ;
    }
